tests: Extract GetCertsForPods (and related functions) from utils
Reduce utils.go file size.
